pyproject.toml
[project]
name = "gcp-terminal-copilot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.6.0",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
]
